camera,object更新
追加
カメラにisObjectFaceCamera関数
objectPartにForward引数
objectPartにUpdateWorldMatrix関数

+++ kEngine/Manager/LightManager/LightManager.cpp
@@ -84,7 +84,7 @@ int LightManager::GetLightCount() {
 	if (lightGPUList_.empty()) {
 		return 1;
 	}
-	return lightGPUList_.size();
+	return (int)lightGPUList_.size();
 }
 
 void LightManager::ConvertDirectionalLight(DirectionalLight* light) {
